restructure project, add nodemon as dev dep
