+++ scripts/research/whatSeatsExist.js
