+++ scripts/scraping/cinestarFunctions.js
