+++ scripts/scraping/consoleProgress.js
