+++ scripts/scraping/getPerformanceDates.js
