+++ server/api/seating.js
