+++ server/public/analytics/chartsScript.js
