+++ server/public/analytics/index.html
