+++ server/public/analytics/prepareData.js
